refactor(agents): extract attempt tool catalog phase
